requirements.txt
streamlit==1.25.0
pytesseract==0.3.10
Pillow==9.3.0
pandas==2.0.3
